# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Aplikasi REST API",
    "description": "Dokumentasi yang dihasilkan dengan <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/pengaduan/": {
      "get": {
        "tags": [
          "Pengaduan"
        ],
        "description": "Endpoint untuk mengambil semua Pengaduan.",
        "responses": {
          "200": {
            "description": "Output API untuk objek Pengaduan",
            "schema": {
              "$ref": "#/definitions/Pengaduan"
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "Judul": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "deskripsi": {
                  "example": "any"
                },
                "tanggal": {
                  "example": "any"
                },
                "notlp": {
                  "example": "any"
                },
                "fotoBase64": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/pengaduan/{id}": {
      "get": {
        "tags": [
          "Pengaduan"
        ],
        "description": "Endpoint untuk mengambil Pengaduan berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Pengaduan",
            "schema": {
              "$ref": "#/definitions/Pengaduan"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pengaduan"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi Pengaduan.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Pengaduan"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Pengaduan",
            "schema": {
              "$ref": "#/definitions/Pengaduan"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Pengaduan"
        ],
        "description": "Endpoint untuk menghapus Pengaduan.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Pengaduan",
            "schema": {
              "$ref": "#/definitions/Pengaduan"
            }
          }
        }
      }
    },
...
    "/petugasksk/": {
      "get": {
        "tags": [
          "PetugasKsk"
        ],
        "description": "Endpoint untuk mengambil semua PetugasKsk.",
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKsk",
            "schema": {
              "$ref": "#/definitions/PetugasKsks"
            }
          }
        }
      },
      "post": {
        "tags": [
          "PetugasKsk"
        ],
        "description": "Endpoint untuk membuat petugasksk baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi PetugasKsk.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PetugasKsk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKsk",
            "schema": {
              "$ref": "#/definitions/PetugasKsk"
            }
          }
        }
      }
    },
    "/petugasksk/{id}": {
      "get": {
        "tags": [
          "PetugasKsk"
        ],
        "description": "Endpoint untuk mengambil PetugasKsk berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKsk",
            "schema": {
              "$ref": "#/definitions/PetugasKsk"
            }
          }
        }
      },
      "put": {
        "tags": [
          "PetugasKsk"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi PetugasKsk.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PetugasKsk"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKsk",
            "schema": {
              "$ref": "#/definitions/PetugasKsk"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PetugasKsk"
        ],
        "description": "Endpoint untuk menghapus PetugasKsk.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKsk",
            "schema": {
              "$ref": "#/definitions/PetugasKsk"
            }
          }
        }
      }
    },
    "/petugaskebakaran/": {
      "get": {
        "tags": [
          "PetugasKebakaran"
        ],
        "description": "Endpoint untuk mengambil semua PetugasKebakaran.",
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKebakaran",
            "schema": {
              "$ref": "#/definitions/PetugasKebakarans"
            }
          }
        }
      },
      "post": {
        "tags": [
          "PetugasKebakaran"
        ],
        "description": "Endpoint untuk membuat petugaskebakaran baru.",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Informasi PetugasKebakaran.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PetugasKebakaran"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKebakaran",
            "schema": {
              "$ref": "#/definitions/PetugasKebakaran"
            }
          }
        }
      }
    },
    "/petugaskebakaran/{id}": {
      "get": {
        "tags": [
          "PetugasKebakaran"
        ],
        "description": "Endpoint untuk mengambil PetugasKebakaran berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKebakaran",
            "schema": {
              "$ref": "#/definitions/PetugasKebakaran"
            }
          }
        }
      },
      "put": {
        "tags": [
          "PetugasKebakaran"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi PetugasKebakaran.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/PetugasKebakaran"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKebakaran",
            "schema": {
              "$ref": "#/definitions/PetugasKebakaran"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "PetugasKebakaran"
        ],
        "description": "Endpoint untuk menghapus PetugasKebakaran.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek PetugasKebakaran",
            "schema": {
              "$ref": "#/definitions/PetugasKebakaran"
            }
          }
        }
      }
    },
...
    "/kebakaran/": {
      "get": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk mengambil semua Kebakaran.",
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "judul": {
                  "example": "any"
                },
                "lokasi": {
                  "example": "any"
                },
                "deskripsi": {
                  "example": "any"
                },
                "tanggal": {
                  "example": "any"
                },
                "notlp": {
                  "example": "any"
                },
                "fotoBase64": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/kebakaran/{id}": {
      "get": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk mengambil Kebakaran berdasarkan id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk index http put.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "data",
            "in": "body",
            "description": "Informasi Kebakaran.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Kebakaran"
        ],
        "description": "Endpoint untuk menghapus Kebakaran.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Output API untuk objek Kebakaran",
            "schema": {
              "$ref": "#/definitions/Kebakaran"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Pengaduan": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "Judul": {
          "type": "string",
          "example": "Banjir Bandang"
        },
        "lokasi": {
          "type": "string",
          "example": "Jl Menjangan IV"
        },
        "deskripsi": {
          "type": "string",
          "example": "Air Tidak Mengalir"
        },
        "tanggal": {
          "type": "string",
          "example": "Selasa 26 Novber 2021"
        },
        "notlp": {
          "type": "string",
          "example": "081387887675"
        },
        "foto": {
          "type": "string",
          "example": "cover.jpg"
        }
      }
    },
...
    "PetugasKebakaran": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "nama": {
          "type": "string",
          "example": "budi"
        },
        "email": {
          "type": "string",
          "example": "budijati@gmail.com"
        },
        "password": {
          "type": "string",
          "example": "12345"
        }
      }
    },
    "PetugasKebakarans": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/PetugasKebaran"
      }
    },
    "Kebakaran": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 1
        },
        "Judul": {
          "type": "string",
          "example": "Banjir Bandang"
        },
        "lokasi": {
          "type": "string",
          "example": "Jl Menjangan IV"
        },
        "deskripsi": {
          "type": "string",
          "example": "Air Tidak Mengalir"
        },
        "tanggal": {
          "type": "string",
          "example": "Selasa 26 Novber 2021"
        },
        "notlp": {
          "type": "string",
          "example": "081387887675"
        },
        "foto": {
          "type": "string",
          "example": "cover.jpg"
        }
      }
    },
...
  }
}